+++ changelog.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.28.0] 2023-08-25
 
-- Add shape.Group
+- Add shape.Group and shape.Container
 - shape.Card default width is 310px
 - shape.Card takes optional note and text